+++ lib/Core/isDefined.ts
@@ -1,3 +1,3 @@
 export default function isDefined<T>(value?: T): value is T {
-  return value !== undefined;
+  return typeof value !== "undefined";
 }
