+++ lib/Models/Catalog/CatalogItems/GeoJsonCatalogItem.ts
@@ -1,23 +1,23 @@
 import i18next from "i18next";
-import { computed, runInAction, toJS } from "mobx";
-import isDefined from "../../../Core/isDefined";
-import JsonValue, { isJsonObject, JsonObject } from "../../../Core/Json";
-import { isZip, parseZipJsonBlob } from "../../../Core/loadBlob";
+import { computed, toJS } from "mobx";
 import {
+  downloadInterruptedByUserError,
   fetchBlob,
   fetchJson,
   isOverMaxSizeResponse,
-  OverMaxSizeResponse
-} from "../../../Core/loadWithProgress";
+  LoadOptions,
+  LoadOptionsWithMaxSize
+} from "../../../Core/fetchWithProgress";
+import isDefined from "../../../Core/isDefined";
+import JsonValue, { isJsonObject, JsonObject } from "../../../Core/Json";
+import { isZip, parseZipJsonBlob } from "../../../Core/loadBlob";
 import readJson from "../../../Core/readJson";
-import TerriaError, { TerriaErrorSeverity } from "../../../Core/TerriaError";
-import { getName } from "../../../ModelMixins/CatalogMemberMixin";
+import TerriaError from "../../../Core/TerriaError";
 import GeoJsonMixin, {
   FeatureCollectionWithCrs,
   toFeatureCollection
 } from "../../../ModelMixins/GeojsonMixin";
 import GeoJsonCatalogItemTraits from "../../../Traits/TraitsClasses/GeoJsonCatalogItemTraits";
-import CommonStrata from "../../Definition/CommonStrata";
 import CreateModel from "../../Definition/CreateModel";
 import Terria from "../../Terria";
 import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
@@ -44,9 +44,7 @@ class GeoJsonCatalogItem extends GeoJsonMixin(
     return isDefined(this._file);
   }
 
-  protected async forceLoadGeojsonData(
-    ignoreMaxFileSize?: true
-  ): Promise<FeatureCollectionWithCrs> {
+  protected async forceLoadGeojsonData(): Promise<FeatureCollectionWithCrs> {
     let jsonData: JsonValue | undefined = undefined;
 
     // GeoJsonCatalogItemTraits.geoJsonData
@@ -71,16 +69,21 @@ class GeoJsonCatalogItem extends GeoJsonMixin(
     else if (this.url) {
       // URL to zipped fle
       const url = proxyCatalogItemUrl(this, this.url);
-      const fetchOptions = {
+
+      if (isZip(url) && typeof FileReader === "undefined") {
+        throw fileApiNotSupportedError(this.terria);
+      }
+
+      const fetchOptions: LoadOptionsWithMaxSize | LoadOptions = {
         bodyObject: this.requestData
           ? (toJS(this.requestData) as JsonObject)
           : undefined,
         asForm: this.postRequestDataAsFormData,
-        maxFileSize: ignoreMaxFileSize ? undefined : 50 * 1024 * 1024
+        fileSizeToPrompt: 50 * 1024 * 1024, // 50 MB,
+        maxFileSize: 250 * 1024 * 1024, // 250 MB,
+        model: this
       };
-      if (isZip(url) && typeof FileReader === "undefined") {
-        throw fileApiNotSupportedError(this.terria);
-      }
+
       const response = (
         await (isZip(url)
           ? fetchBlob(url, fetchOptions)
@@ -88,7 +91,8 @@ class GeoJsonCatalogItem extends GeoJsonMixin(
       ).throwIfUndefined("Failed to download GeoJSON");
 
       if (isOverMaxSizeResponse<Blob | JsonValue>(response)) {
-        return await this.largeDownloadWarning(response);
+        this.resetLoadMapItems();
+        throw downloadInterruptedByUserError(response);
       } else {
         if (response.response instanceof Blob) {
           jsonData = await parseZipJsonBlob(response.response);
@@ -108,47 +112,6 @@ class GeoJsonCatalogItem extends GeoJsonMixin(
     }
 
     throw TerriaError.from("Failed to load geojson");
-  }
-
-  private async largeDownloadWarning(response: OverMaxSizeResponse) {
-    runInAction(() =>
-      this.setTrait(CommonStrata.underride, "disablePreview", true)
-    );
-
-    const fileSizeMessage =
-      response.overMaxFileSize.type === "total"
-        ? `The size of the file is ${Math.round(
-            response.overMaxFileSize.bytes / (1024 * 1024)
-          )}MB`
-        : `The size of the file is over ${Math.round(
-            response.overMaxFileSize.bytes / (1024 * 1024)
-          )}MB`;
-
-    const result = await new Promise<boolean>(resolve => {
-      this.terria.notificationState.addNotificationToQueue({
-        title: "Warning: large file download",
-        message: `You are about to download a large file for item: \`${getName(
-          this
-        )}\`\n${fileSizeMessage}`,
-        confirmAction: () => resolve(true),
-        denyAction: () => resolve(false),
-        denyText: "Cancel",
-        confirmText: "Download"
-      });
-    });
-
-    if (result) {
-      return this.forceLoadGeojsonData(true);
-    } else {
-      this.resetLoadMapItems();
-      throw new TerriaError({
-        title: "File download interrupted by user",
-        message: fileSizeMessage,
-        overrideRaiseToUser: false,
-        importance: 3,
-        severity: TerriaErrorSeverity.Error
-      });
-    }
   }
 }
